+++ apps/web/src/hooks/queryGetCurrentUser.graphql
@@ -11,5 +11,6 @@ query GetCurrentUser {
     location
     website
     twitter
+    
   }
 }
