--- apps/web/src/pages/User/ProfileInfo/index.ts
@@ -1,3 +1,0 @@
-import ProfileInfo from './ProfileInfo';
-
-export default ProfileInfo;
