+++ apps/web/src/pages/User/UserInfo/index.ts
@@ -1,0 +1,3 @@
+import UserInfo from './UserInfo';
+
+export default UserInfo;
