+++ apps/web/src/pages/User/queryUserProjects.graphql
@@ -14,5 +14,6 @@ query GetUserProjects($userId: String!, $input: SearchProjectsInput) {
         avatar
       }
     }
+    totalCount
   }
 }
